backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
fastapi==0.115.5
uvicorn==0.41.0
pydantic==2.12.5
sqlmodel>=0.0.22
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
google-generativeai>=0.8.0
httpx>=0.27.0
